UI improvements
Corrected arc size, at first I kinda liked that roundness does not match, but now I think it is much better when it does

Also fixed painting background with rounded corners by using `Graphics2D.fill(Shape s)`. The reason to use this instead of `Graphics.fillRoundRect` is that the latter has an off-by-one error that is still not fixed. Due to this bottom corners are not painted correctly